+++ src/services/chaintracker/chaintracks/__tests/ChaintracksClientApi.test.ts
@@ -22,13 +22,14 @@ clientClass = 'Chaintracks'
 const includeLocalServiceClient = true
 const includeLocalServiceChaintracks = true
 const includeNpmRegistryClient = true
+const includeGcrTestClient = true
 
 describe(`ChaintracksClientApi tests`, () => {
   jest.setTimeout(999999999)
 
   const chain: Chain = 'main'
 
-  const clients: ChaintracksClientApi[] = []
+  const clients: { client: ChaintracksClientApi; chain: Chain }[] = []
 
   let localService: ChaintracksService
   let localServiceStorage: ChaintracksStorageApi
@@ -43,19 +44,28 @@ describe(`ChaintracksClientApi tests`, () => {
 
       if (includeLocalServiceClient) {
         localServiceClient = new ChaintracksServiceClient(chain, `http://localhost:${localService.port}`, {})
-        clients.push(localServiceClient)
+        clients.push({ client: localServiceClient, chain })
       }
 
       if (includeLocalServiceChaintracks) {
-        clients.push(localService.chaintracks)
+        clients.push({ client: localService.chaintracks, chain })
       }
     }
 
-    if (includeNpmRegistryClient) {
-      clients.push(makeNpmRegistryClient(chain))
+    if (includeGcrTestClient) {
+      const gcr = new ChaintracksServiceClient('test', `https://testnet-chaintracks.babbage.systems`, {})
+      clients.push({ client: gcr, chain: 'test' })
+    }
+    if (includeGcrTestClient) {
+      const gcr = new ChaintracksServiceClient('main', `https://mainnet-chaintracks.babbage.systems`, {})
+      clients.push({ client: gcr, chain: 'main' })
     }
 
-    const ft = await clients[0].findChainTipHeader()
+    if (includeNpmRegistryClient) {
+      clients.push({ client: makeNpmRegistryClient(chain), chain })
+    }
+
+    const ft = await clients[0].client.findChainTipHeader()
     if (!ft) throw new Error('No chain tip found')
     firstTip = ft
   })
@@ -65,30 +75,30 @@ describe(`ChaintracksClientApi tests`, () => {
   })
 
   test('0 getChain', async () => {
-    for (const client of clients) {
+    for (const { client, chain } of clients) {
       const gotChain = await client.getChain()
       expect(gotChain).toBe(chain)
     }
   })
 
   test('1 getInfo', async () => {
-    for (const client of clients) {
+    for (const { client, chain } of clients) {
       const gotInfo = await client.getInfo()
       expect(gotInfo.chain).toBe(chain)
       expect(gotInfo.heightBulk).toBeGreaterThan(700000)
-      expect(gotInfo.heightLive).toBeGreaterThanOrEqual(firstTip.height)
+      expect(gotInfo.heightLive).toBeGreaterThanOrEqual(firstTip.height - 2)
     }
   })
 
   test('2 getPresentHeight', async () => {
-    for (const client of clients) {
+    for (const { client, chain } of clients) {
       const presentHeight = await client.getPresentHeight()
-      expect(presentHeight).toBeGreaterThanOrEqual(firstTip.height)
+      expect(presentHeight).toBeGreaterThanOrEqual(firstTip.height - 2)
     }
   })
 
   test('3 getHeaders', async () => {
-    for (const client of clients) {
+    for (const { client, chain } of clients) {
       const info = await client.getInfo()
       const h0 = info.heightBulk + 1
       const h1 = info.heightLive || 10
@@ -113,21 +123,21 @@ describe(`ChaintracksClientApi tests`, () => {
   })
 
   test('4 findChainTipHeader', async () => {
-    for (const client of clients) {
+    for (const { client, chain } of clients) {
       const tipHeader = await client.findChainTipHeader()
       expect(tipHeader.height >= firstTip.height).toBe(true)
     }
   })
 
   test('5 findChainTipHash', async () => {
-    for (const client of clients) {
+    for (const { client, chain } of clients) {
       const hash = await client.findChainTipHash()
       expect(hash.length === 64).toBe(true)
     }
   })
 
   test('6 findHeaderForHeight', async () => {
-    for (const client of clients) {
+    for (const { client, chain } of clients) {
       const header0 = await client.findHeaderForHeight(0)
       expect(header0 !== undefined).toBe(true)
       if (header0) {
@@ -137,13 +147,13 @@ describe(`ChaintracksClientApi tests`, () => {
       const header = await client.findHeaderForHeight(firstTip.height)
       expect(header && header.height === firstTip.height).toBe(true)
 
-      const missing = await client.findHeaderForHeight(1000 + firstTip.height)
+      const missing = await client.findHeaderForHeight(99999999)
       expect(missing === undefined).toBe(true)
     }
   })
 
   test('7 addHeader', async () => {
-    for (const client of clients) {
+    for (const { client, chain } of clients) {
       const t = await client.findChainTipHeader()
       const h: BaseBlockHeader = {
         version: t.version,