+++ src/utility/utilityHelpers.ts
@@ -151,9 +151,16 @@ export function verifyOne<T>(results: T[], errorDescrition?: string): T {
 
 /**
  * Returns an await'able Promise that resolves in the given number of msecs.
+ * @param msecs number of milliseconds to wait before resolving the promise.
+ * Must be greater than zero and less than 2 minutes (120,000 msecs)
  * @publicbody
  */
 export function wait(msecs: number): Promise<void> {
+  const MIN_WAIT = 0
+  const MAX_WAIT = 2 * 60 * 1000 // maximum allowed wait in ms (2 minutes)
+  if (typeof msecs !== 'number' || !Number.isFinite(msecs) || isNaN(msecs) || msecs < MIN_WAIT || msecs > MAX_WAIT) {
+    throw new WERR_INVALID_PARAMETER('msecs', `a number between ${MIN_WAIT} and ${MAX_WAIT} msecs, not ${msecs}.`)
+  }
   return new Promise(resolve => setTimeout(resolve, msecs))
 }